# openapi.json
{
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "definitions": {
...
    "ErrorResponse": {
      "properties": {
        "error": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
    "VoiceAudioUploadResponse": {
      "properties": {
        "data": {
          "properties": {
            "audio_id": {
              "type": "string"
            },
            "filename": {
              "type": "string"
            },
            "processing_status": {
              "type": "string"
            },
            "size": {
              "type": "integer"
            },
            "status": {
              "type": "string"
            },
            "uploaded_at": {
              "format": "date-time",
              "type": "string"
            }
          },
          "type": "object"
        },
        "message": {
          "type": "string"
        },
        "success": {
          "example": true,
          "type": "boolean"
        }
      },
      "type": "object"
    },
    "VoiceNoteResponse": {
      "properties": {
        "data": {
          "properties": {
            "category": {
              "type": "string"
            },
            "created_at": {
              "format": "date-time",
              "type": "string"
            },
            "note_id": {
              "type": "string"
            },
            "priority": {
              "type": "string"
            },
            "processed": {
              "type": "boolean"
            },
            "tags": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "transcript": {
              "type": "string"
            }
          },
          "type": "object"
        },
        "message": {
          "type": "string"
        },
        "success": {
          "example": true,
          "type": "boolean"
        }
      },
      "type": "object"
    }
  },
  "info": {
    "contact": {
      "name": "RouteForce Support",
      "url": "https://github.com/ApacheEcho/RouteForceRouting"
    },
    "description": "API documentation for RouteForce Routing application - A comprehensive route optimization platform for field execution teams",
    "title": "RouteForce Routing API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/voice/audio/upload": {
      "post": {
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "description": "Audio file to upload (wav, mp3, ogg, webm, m4a)",
            "in": "formData",
            "name": "audio",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Audio uploaded",
            "schema": {
              "$ref": "#/definitions/VoiceAudioUploadResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Handle audio file upload for offline voice processing",
        "tags": [
          "Voice"
        ]
      }
    },
    "/api/voice/code-generation": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "component_name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "function_name": {
                  "type": "string"
                },
                "language": {
                  "type": "string"
                }
              },
              "required": [
                "description"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code generated",
            "examples": {
              "application/json": {
                "data": {
                  "code": "def new_function():\n    # calculate something from input\n    pass\n",
                  "description": "calculate something from input",
                  "generated_at": "2024-01-15T10:30:00Z",
                  "language": "python",
                  "metadata": {
                    "characters": 120,
                    "function_name": "new_function",
                    "lines": 6
                  }
                },
                "success": true
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Generate code from voice input description",
        "tags": [
          "Voice"
        ]
      }
    },
    "/api/voice/commit": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "author": {
                  "type": "string"
                },
                "files": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "message"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commit accepted"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Handle voice-dictated commit messages",
        "tags": [
          "Voice"
        ]
      }
    },
    "/api/voice/notes": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "category": {
                  "type": "string"
                },
                "priority": {
                  "type": "string"
                },
                "tags": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "transcript": {
                  "type": "string"
                }
              },
              "required": [
                "transcript"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voice note saved",
            "schema": {
              "$ref": "#/definitions/VoiceNoteResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Save voice note for later processing",
        "tags": [
          "Voice"
        ]
      }
    },
    "/api/voice/settings": {
      "get": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "properties": {
                "auto_save": {
                  "type": "boolean"
                },
                "language": {
                  "type": "string"
                },
                "noise_reduction": {
                  "type": "boolean"
                },
                "push_to_talk": {
                  "type": "boolean"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settings retrieved/updated"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get or update voice settings for the user",
        "tags": [
          "Voice"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "properties": {
                "auto_save": {
                  "type": "boolean"
                },
                "language": {
                  "type": "string"
                },
                "noise_reduction": {
                  "type": "boolean"
                },
                "push_to_talk": {
                  "type": "boolean"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settings retrieved/updated"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get or update voice settings for the user",
        "tags": [
          "Voice"
        ]
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "in": "header",
      "name": "X-API-Key",
      "type": "apiKey"
    },
    "JWT": {
      "description": "JWT Bearer token. Use: 'Bearer <token>'",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}